pyproject.toml
[tool.poetry]
name = "fastapi_auth"
version = "0.1.0"
description = ""
authors = ["Felipe Curcio <felipe_nieto010@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.7.0"
python-multipart = "^0.0.9"
bcrypt = "^4.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.scripts]
start = 'dale.index:main'

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 101

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # flake8
    "I", # isort
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
]


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true